Revert renaming class MasterService and FakeThreadPoolMasterService
