+++ server/src/main/java/org/opensearch/common/util/concurrent/BaseFuture.java
@@ -33,7 +33,7 @@
 package org.opensearch.common.util.concurrent;
 
 import org.opensearch.cluster.service.ClusterApplierService;
-import org.opensearch.cluster.service.ClusterManagerService;
+import org.opensearch.cluster.service.MasterService;
 import org.opensearch.common.Nullable;
 import org.opensearch.threadpool.ThreadPool;
 import org.opensearch.transport.Transports;
@@ -109,7 +109,7 @@ public abstract class BaseFuture<V> implements Future<V> {
         return Transports.assertNotTransportThread(BLOCKING_OP_REASON)
             && ThreadPool.assertNotScheduleThread(BLOCKING_OP_REASON)
             && ClusterApplierService.assertNotClusterStateUpdateThread(BLOCKING_OP_REASON)
-            && ClusterManagerService.assertNotMasterUpdateThread(BLOCKING_OP_REASON);
+            && MasterService.assertNotMasterUpdateThread(BLOCKING_OP_REASON);
     }
 
     @Override
