+++ server/src/test/java/org/opensearch/snapshots/SnapshotResiliencyTests.java
@@ -144,8 +144,8 @@ import org.opensearch.cluster.routing.allocation.AllocationService;
 import org.opensearch.cluster.routing.allocation.command.AllocateEmptyPrimaryAllocationCommand;
 import org.opensearch.cluster.service.ClusterApplierService;
 import org.opensearch.cluster.service.ClusterService;
-import org.opensearch.cluster.service.FakeThreadPoolClusterManagerService;
-import org.opensearch.cluster.service.ClusterManagerService;
+import org.opensearch.cluster.service.FakeThreadPoolMasterService;
+import org.opensearch.cluster.service.MasterService;
 import org.opensearch.common.CheckedConsumer;
 import org.opensearch.common.Nullable;
 import org.opensearch.common.io.stream.NamedWriteableRegistry;
@@ -1643,7 +1643,7 @@ public class SnapshotResiliencyTests extends OpenSearchTestCase {
 
             private final DiscoveryNode node;
 
-            private final ClusterManagerService clusterManagerService;
+            private final MasterService clusterManagerService;
 
             private final AllocationService allocationService;
 
@@ -1663,7 +1663,7 @@ public class SnapshotResiliencyTests extends OpenSearchTestCase {
                 this.node = node;
                 final Environment environment = createEnvironment(node.getName());
                 threadPool = deterministicTaskQueue.getThreadPool(runnable -> CoordinatorTests.onNodeLog(node, runnable));
-                clusterManagerService = new FakeThreadPoolClusterManagerService(
+                clusterManagerService = new FakeThreadPoolMasterService(
                     node.getName(),
                     "test",
                     threadPool,
